Add trait for asserting schema requests
By default we can extend provided test case, for situations where we can't extend the test case we can use trait AssertRequestAgainstSchema.
